Tweaked an issue with an upgraded compiler version on MacOS X that was impacted by passing the C++ compiler a C-centric command line flag; which now throws an error refusing to compile.
The C flag was previously added to work-around a CUDA header file
error on OS X. This header file now appears to be fixed in the latest
release of CUDA (4.1) -- everyone (at least on MacOS X) should update
their version of CUDA as it has several other bug fixes.

This commit removes the C language compile flag from the build process
(which is more appropriate for us with NVIDIA having patched the
header file).