A significant set of overhauls/refeactoring and a few bug fixes (that appear to overlap with Dean's last set of commits).
Overall this is a focus on separating scout-centric types from clang
provided types and how it relates to parsing.  As a first step this
version removes handling almost all 'forall' constructs within one
primary interfaces and instead specializes based on the statement
types.  For example, this commit has support for parsing 'forall mesh'
statements -- as such, other forall types (renderall and array-based
versions) are disabled (commented out).  These additional types need
to be reimplemented independently...

This version is still choking in the code gen (and I think some of the
code gen pieces are totally broken beyond that).  Based on the changes
made in parsing and the introduction of new (more specialized) types
it will likely be best to refactor all the codegen pieces as well.
The primary goal should be getting sequential code generation back into
shape and ignore both parallelism and GPU code gen for starters...

Language changes: note that we now use 'in' vs. 'of' in forall mesh
syntax (I have not yet updated the regression tests to account for
this change).

I worked on cleaning up some error messages and naming conventions
within the parsing code.