A significant set of overhaul work:
  (1) Rework of mesh and mesh field decls, types, etc.  Many
      scout-centric types are now stand alone vs. inherited and thus
      have isolated implementations.

  (2) Start of parsing code now refactor to address specific types
      (e.g. parse forall mesh, parse renderall, parse forall-array are
      now designed to be separted vs. itegrated into one routine).

  (3) New error cases handled, improved error messages (likely a lot
      more work here).

  (4) Start of refactoring parts of the runtime and standard library
      to allow a better mix of Scout and C/C++ source.  Real routines
      in standard library vs. header file only implementations.

NOTES: renderall and forall-array functionality is currently commented
out to reduce complexity in the refactoring. A lot of clean up is
likely needed to further separate scout pieces from the main clang
base code.  We also need to re-sync up with the main clang repository
before we drift too far from the base.