Some clean up of duplicated files (in git-tools and tools).
Make a directory for the sublime editor files.

WIP: Reworked command line arg for not including the standard library
automatically 'noscstdlib'.  Unlike our past use of command line
arguments, this one is now moved internally within Driver and is
implemented to match the internal models in clang (which appear to
have changed with some of the latest code in the trunk).