Squashed 'llvm/tools/clang/' changes from d22eccf..e3c4307
e3c4307 D9600: Add scan-build python implementation
09b896f [CUDA] Add explicit mapping from sm_XX to compute_YY.
03e92f7 Improve AST dumping:
89dda38 Properly track the end location of an exception specification.
39fe6b9 Module debugging: Make the module format part of the module hash instead of the file name. This is consistent with how other HeaderSearchOptions are handled.
d7aab06 Don't store CGOpenMPRegionInfo::CodeGen as a reference (PR26078)
d83e5d4 [modules] Don't diagnose a conflict between two using-declarations that name equivalent internal linkage entities.
6e332e0 Add a clang test for r257376 (Ensure -mcpu=xscale works for arm targets).
523c668 function names start with a lower case letter ; NFC
0e2452f Revert r253898 and r253899, this breaks mingw compilation on openSUSE. Will re-apply after llvm 3.8 is branched.
8b94c48 [analyzer] Fix RangeConstraintManager's pinning of single value ranges.
8c0250d [analyzer] Evaluate integral casts as cast symbols if truncations are detected.
81bc392 PR18513: make gcc compatible layout for bit-fields with explicit aligned attribute, by Dmitry Polukhin Fix binary compatibility issue with GCC. Differential Revision: http://reviews.llvm.org/D14980
70e761e Fix infinite recursion for invalid declaration, by Dmitry Polukhin Fix for a case found by fuzzing PR23057 (comment #25 https://llvm.org/bugs/show_bug.cgi?id=23057#c25). Differential Revision: http://reviews.llvm.org/D16065
318abdb clang-format: [JS] Support exporting abstract classes.
a791c7f [CUDA] Reject values for --cuda-gpu-arch that are not of the form /sm_\d+/.
d168c09 Add an Action* member to InputInfo.
97672aa Make Driver::BuildJobsForAction return an InputInfo, instead of using an outparam.
089f0f9 Move ownership of Action objects into Compilation.
8e1a761 clang-format: [JS] Teach clang-format about "export interface".
a454e26 When a tag is declared in prototype scope in C, if we've decided that it redeclares an existing tag but are creating a new declaration anyway (because it has attributes or changes the visibility of the name), don't warn that it won't be visible outside the current scope. That's not true.

git-subtree-dir: llvm/tools/clang
git-subtree-split: e3c4307458824602f2ec0703575353b445941b01